Updating underlying datatypes used for Vectors in some places. Changed the enum/enum class being used

+++ Keyboard.cpp
@@ -4,8 +4,8 @@ using namespace NCL;
 
 Keyboard::Keyboard() {
 	//Initialise the arrays to false!
-	memset(keyStates , 0, (int)KeyCodes::MAXVALUE * sizeof(bool));
-	memset(holdStates, 0, (int)KeyCodes::MAXVALUE * sizeof(bool));
+	memset(keyStates , 0, KeyCodes::MAXVALUE * sizeof(bool));
+	memset(holdStates, 0, KeyCodes::MAXVALUE * sizeof(bool));
 }
 
 /*
@@ -13,13 +13,13 @@ Updates variables controlling whether a keyboard key has been
 held for multiple frames.
 */
 void Keyboard::UpdateFrameState(float msec) {
-	memcpy(holdStates, keyStates, (int)KeyCodes::MAXVALUE * sizeof(bool));
+	memcpy(holdStates, keyStates, KeyCodes::MAXVALUE * sizeof(bool));
 }
 
 void Keyboard::Sleep() {
 	isAwake = false; 
-	memset(keyStates , 0, (int)KeyCodes::MAXVALUE * sizeof(bool));
-	memset(holdStates, 0, (int)KeyCodes::MAXVALUE * sizeof(bool));
+	memset(keyStates , 0, KeyCodes::MAXVALUE * sizeof(bool));
+	memset(holdStates, 0, KeyCodes::MAXVALUE * sizeof(bool));
 }
 
 void Keyboard::Wake() {
